+++ java/runtime/src/main/java/org/ray/runtime/AbstractRayRuntime.java
@@ -24,7 +24,7 @@ import org.ray.runtime.objectstore.ObjectStoreProxy;
 import org.ray.runtime.objectstore.ObjectStoreProxy.GetStatus;
 import org.ray.runtime.raylet.RayletClient;
 import org.ray.runtime.task.TaskSpec;
-import org.ray.runtime.util.ArgumentsBuilder;
+import org.ray.runtime.task.ArgumentsBuilder;
 import org.ray.runtime.util.UniqueIdHelper;
 import org.ray.runtime.util.MethodId;
 import org.ray.runtime.util.ResourceUtil;
